+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.3.0 - June 24, 2013
 
 - Added company_code to Avalara::Configuration to be used as a default for request invoice and request cancel_tax